We use a+ agin for writing the file. There is a problem here.
The file has data from several parts of the profile: srcips and dstip
Sometimes we have more data from them after some TW. Which means that we
want to append data. But the original data is still there, so the data
is repeated. And we can not use w+ because when we put the srcip we
delete the dstip