profilerProcess.py: __database__.add_out_srcport()
We do not care about srcPort in case of traffic going out.
In same way in case of traffic going in we do not care about dstPort.

profilerProcess.py: elif self.analysis_direction == 'all':

The possibility "flow is going from homenet to homenet" was added.
Also it was re-structured for better understanding.