logsProcess.py
1. On the beginning I put the timeWindow ID  because It seems to me more understandable to know what is "my" IP during the analyzing the file.
Of course in many editors you can see name of the file during the seeing into the file but not all of them.

2. Information about blocking and evidence were put in the begginning. As for me it is the most important information if I analyzing my traffic.

3. Dst ports and src ports.
These functions are put in a loop. I am not sure if it is clear enough (as for me yes). The reason why I did it is that the differences between the src and dst ports code blocks are insignificant. If the text should be more different for each port I will put inside 2 code blocks instead of for loop.