progress_bar.py
It is just testing functionality, because when malicious IPs are loaded in Threat inteligence module, it takes some seconds (6-7s). So I think that user should know that something is happening. It seems to me better than just no write something like : "Loading Ips.". However it is up to you. Try to run slips and tell me if you like such bar.