slips.py
I put loading modules on the beginning, because for example Threat inteligence module needs 6-7s to load all malicious IPs. So id slips run and also loading malicious ips run as well, so we could miss some malicious ips. So it seems to me better to wait for all modules and then start the rest of the process.