profileProcess.py
Creating a sequence of symbols is done. It is not tested and probably there is misstake (I am not sure about the zeros on the beggining etc.)
So I will test it on data provided by you.

Instead of storing "last_ts" and "T2" I store "last_last_ts" and "last_ts". It seems better for understanding.