slips.py: Add Frantisek Strasak changes
A checking of available redis server and bro is added.

Importing of modules was moved to main function. The reason is that
if something is not installed or not ready (for example redis server is not up)
so the user still should run the slips for help at least without getting error.

def test_redis_database()
Function to check if redis server is up.

def test_zeek()
Function to check if bro is runable. It is tested by command "bro --version".

def terminate_slips()
Function to manage terminating of slips. If in the future some deleting files will be needed,
it will be inside this function.