profilerProcess.py
The defining of time format was wrong (Frenky's bad). So now for each type of input (suricata, zeek, argus) the time is converted to datetime.datetime object. Of course if there is no information about time format in condif file we have to detect it. There are some classic time format for detection for each input.